+++ example/calc_parser.dart
@@ -561,7 +561,7 @@ class CalcParser {
         // !.
         state.ok = state.pos >= state.input.length;
         if (!state.ok) {
-          state.fail(const ErrorExpectedEndOfInput());
+          state.fail(const ErrorUnexpectedCharacter());
         }
         if (state.ok) {
           $0 = $2;