+++ lib/src/expression_generators/not_predicate_generator.dart
@@ -11,12 +11,6 @@ if (!state.ok) {
   state.pos = {{pos}};
 }''';
 
-  static const _templateEof = '''
-state.ok = state.pos >= state.input.length;
-if (!state.ok) {
-  state.fail(const ErrorExpectedEndOfInput());
-}''';
-
   NotPredicateGenerator({
     required super.expression,
     required super.ruleGenerator,
@@ -24,23 +18,48 @@ if (!state.ok) {
 
   @override
   String generate() {
-    final child = expression.expression;
-    if (child is AnyCharacterExpression) {
-      return _generateEof();
+    final optimized = _NotPredicateGenerator2.optimize(this);
+    if (optimized != null) {
+      return optimized;
     }
 
-    return _generate();
-  }
-
-  String _generate() {
     final values = <String, String>{};
     final child = expression.expression;
     values['pos'] = allocateName();
     values['p'] = generateExpression(child, false);
     return render(_template, values);
   }
+}
 
-  String _generateEof() {
-    return render(_templateEof, {});
+class _NotPredicateGenerator2
+    extends ExpressionGenerator<NotPredicateExpression> {
+  static const _template = '''
+state.ok = state.pos >= state.input.length;
+if (!state.ok) {
+  state.fail(const ErrorUnexpectedCharacter());
+}''';
+
+  _NotPredicateGenerator2({
+    required super.expression,
+    required super.ruleGenerator,
+  });
+
+  @override
+  String generate() {
+    return render(_template, const {});
+  }
+
+  static String? optimize(NotPredicateGenerator generator) {
+    final expression = generator.expression;
+    final child = expression.expression;
+    if (child is! AnyCharacterExpression) {
+      return null;
+    }
+
+    final generator2 = _NotPredicateGenerator2(
+      expression: expression,
+      ruleGenerator: generator.ruleGenerator,
+    );
+    return generator2.generate();
   }
 }